Move shared job serialization and backoff helpers to BaseAdapter
